+++ src/quick-lint-js/fe/lint.cpp
@@ -239,9 +239,7 @@ void linter::visit_enter_namespace_scope() {
   QLJS_UNIMPLEMENTED();  // TODO(#690)
 }
 
-void linter::visit_enter_type_alias_scope() {
-  QLJS_UNIMPLEMENTED();  // TODO(#690)
-}
+void linter::visit_enter_type_alias_scope() { this->scopes_.push(); }
 
 void linter::visit_exit_block_scope() {
   QLJS_ASSERT(!this->scopes_.empty());
@@ -320,7 +318,11 @@ void linter::visit_exit_namespace_scope() {
 }
 
 void linter::visit_exit_type_alias_scope() {
-  QLJS_UNIMPLEMENTED();  // TODO(#690)
+  QLJS_ASSERT(!this->scopes_.empty());
+  this->propagate_variable_uses_to_parent_scope(
+      /*allow_variable_use_before_declaration=*/false,
+      /*consume_arguments=*/false);
+  this->scopes_.pop();
 }
 
 void linter::visit_keyword_variable_use(identifier) {