+++ test/test-lint.cpp
@@ -2873,6 +2873,100 @@ TEST(test_lint_interface,
   }
 }
 
+TEST(test_lint_type_alias, type_alias_can_use_outside_types) {
+  const char8 imported_declaration[] = u8"C";
+  const char8 type_alias_declaration[] = u8"Alias";
+  const char8 type_use[] = u8"C";
+
+  {
+    // import {C} from "other-module";
+    // type Alias = C;
+    diag_collector v;
+    linter l(&v, &default_globals);
+    l.visit_variable_declaration(identifier_of(imported_declaration),
+                                 variable_kind::_import,
+                                 variable_init_kind::normal);
+    l.visit_variable_declaration(identifier_of(type_alias_declaration),
+                                 variable_kind::_type_alias,
+                                 variable_init_kind::normal);
+    l.visit_enter_type_alias_scope();
+    l.visit_variable_type_use(identifier_of(type_use));
+    l.visit_exit_type_alias_scope();
+    l.visit_end_of_module();
+
+    EXPECT_THAT(v.errors, IsEmpty());
+  }
+
+  {
+    // type Alias = C;  // ERROR
+    diag_collector v;
+    linter l(&v, &default_globals);
+    l.visit_variable_declaration(identifier_of(type_alias_declaration),
+                                 variable_kind::_type_alias,
+                                 variable_init_kind::normal);
+    l.visit_enter_type_alias_scope();
+    l.visit_variable_type_use(identifier_of(type_use));
+    l.visit_exit_type_alias_scope();
+    l.visit_end_of_module();
+
+    EXPECT_THAT(v.errors,
+                ElementsAre(DIAG_TYPE_SPAN(diag_use_of_undeclared_type, name,
+                                           span_of(type_use))));
+  }
+}
+
+TEST(test_lint_type_alias, generic_type_alias_parameters_are_usable_inside) {
+  const char8 type_alias_declaration[] = u8"Alias";
+  const char8 parameter_declaration[] = u8"T";
+  const char8 parameter_use[] = u8"T";
+
+  {
+    // type Alias<T> = T;
+    diag_collector v;
+    linter l(&v, &default_globals);
+    l.visit_variable_declaration(identifier_of(type_alias_declaration),
+                                 variable_kind::_type_alias,
+                                 variable_init_kind::normal);
+    l.visit_enter_type_alias_scope();
+    l.visit_variable_declaration(identifier_of(parameter_declaration),
+                                 variable_kind::_generic_parameter,
+                                 variable_init_kind::normal);
+    l.visit_variable_type_use(identifier_of(parameter_use));
+    l.visit_exit_type_alias_scope();
+    l.visit_end_of_module();
+
+    EXPECT_THAT(v.errors, IsEmpty());
+  }
+}
+
+TEST(test_lint_type_alias,
+     generic_type_alias_parameters_are_not_usable_outside) {
+  const char8 type_alias_declaration[] = u8"Alias";
+  const char8 parameter_declaration[] = u8"T";
+  const char8 parameter_use_outside_type_alias[] = u8"T";
+
+  {
+    // type Alias<T> = null;
+    // (null as T);           // ERROR
+    diag_collector v;
+    linter l(&v, &default_globals);
+    l.visit_variable_declaration(identifier_of(type_alias_declaration),
+                                 variable_kind::_type_alias,
+                                 variable_init_kind::normal);
+    l.visit_enter_type_alias_scope();
+    l.visit_variable_declaration(identifier_of(parameter_declaration),
+                                 variable_kind::_generic_parameter,
+                                 variable_init_kind::normal);
+    l.visit_exit_type_alias_scope();
+    l.visit_variable_type_use(identifier_of(parameter_use_outside_type_alias));
+    l.visit_end_of_module();
+
+    EXPECT_THAT(v.errors, ElementsAre(DIAG_TYPE_SPAN(
+                              diag_use_of_undeclared_type, name,
+                              span_of(parameter_use_outside_type_alias))));
+  }
+}
+
 TEST(test_lint_magic_arguments,
      arguments_magic_variable_is_usable_within_functions) {
   const char8 arguments_use[] = u8"arguments";